requirements.txt
# Core
python-dotenv>=1.0
httpx>=0.27

# Blockchain / Polymarket
web3>=6.11
py-clob-client>=0.17
py-order-utils>=0.3

# AI agents
# (XAI/Grok, Anthropic/Claude, DeepSeek are called via httpx REST — no SDK needed)

# Database
# sqlite3 is built-in

# WebSocket (observer)
websocket-client>=1.8

# CLI
typer>=0.12
devtools>=0.12

# ChromaDB (similar-analysis lookup)
chromadb>=0.5
langchain-openai>=0.1
langchain-community>=0.2
langchain-chroma>=0.1

# Optional: external research (Scout agent)
# tavily-python>=0.3
# newsapi-python>=0.2
